Changes
Moved previously existing files in the directory "justice",
Add new files into the root of the directory.
Updated the README file to reflect changes.

Each function from these changes is contained in a file of it's own.
Also a precompiled 'monty' file is added; You can run some tests on it.

Edits to be made:
	Instead of using printf to print the error messages, from your
	existing files I observed that a better implementation would be
	to use fprintf to print to stderr

Please go through these changes, and share ideas with me.